backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg[binary]
python-dotenv
pytest
pytest-cov
coverage
